pyproject.toml
[project]
name = "dspy-compounding-engineering"
version = "0.1.2.2"
description = "DSPy implementation of the Compounding Engineering"
readme = "README.md"
authors = [
    {name = "Strategic Automation", email = "hello@strategicautomation.co.uk"}
]
license = {text = "MIT"}
keywords = ["ai", "llm", "dspy", "compounding", "engineering", "cli", "agent"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.11"
dependencies = [
    "dspy-ai",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "python-dotenv",
    "openai",
    "python-frontmatter",
    "jsonschema",
    "pyyaml>=6.0",
    "filelock>=3.15.4",
    "qdrant-client>=1.7.0",
    "pygments<2.19.0",
    "fastembed>=0.7.4",
    "tiktoken>=0.7.0",
    "beautifulsoup4>=4.14.3",
    "markdownify>=1.2.2",
    "httpx>=0.28.1",
    "loguru>=0.7.2",
    "langfuse>=3.11.2",
    "openinference-instrumentation-dspy>=0.1.32",
    "ruff>=0.14.10",
    "fastmcp>=2.14.2",
    "dspy>=3.0.4",
    "pydantic>=2.12.5",
    "redis>=7.1.0",
    "celery>=5.6.1",
    "dspy-qdrant>=0.1.4",
    "sqlmodel>=0.0.31",
    "networkx[default]>=3.6.1",
    "types-networkx==3.6.1.20251220",
    "mem0ai>=1.0.1",
    "fastapi[standard]>=0.128.0",
    "pydocket>=0.16.3",
    "pyrefly>=0.46.3",
]


# Optional dependencies for full agent functionality
[project.optional-dependencies]
full = [
    "dspy-ai",
    "python-dotenv",
    "openai",
    "python-frontmatter",
    "jsonschema",
    "pyyaml>=6.0",
    "filelock>=3.15.4",
]

[project.scripts]
ce = "cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["agents", "workflows", "utils", "server", "cli.py", "config.py"]

[dependency-groups]
dev = [
    "ruff>=0.14.7",
    "pre-commit>=3.6.0",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "TID", # 👈 import discipline
]
ignore = [
   ## "TID252", # ❌ TEMPORARY — see next section
   ## "C901",
]

[tool.ruff.lint.isort]
known-first-party = ["agents", "workflows", "utils", "server"]

[tool.ruff.lint.per-file-ignores]
"agents/**" = ["TID252"]
"utils/**" = ["TID252"]
"workflows/**" = ["TID252"]
